Replaced usage of non-ANSI sqrtf() with sqrt().

+++ src/rgb.h
@@ -82,9 +82,9 @@ H_INLINE int MMRGBColorSimilarToColor(MMRGBColor c1, MMRGBColor c2,
 		uint8_t d1 = c1.red - c2.red;
 		uint8_t d2 = c1.green - c2.green;
 		uint8_t d3 = c1.blue - c2.blue;
-		return sqrtf((d1 * d1) +
-		             (d2 * d2) +
-		             (d3 * d3)) <= (tolerance * 442.0f);
+		return sqrt((d1 * d1) +
+		            (d2 * d2) +
+		            (d3 * d3)) <= (tolerance * 442.0f);
 	}
 
 }
@@ -98,9 +98,9 @@ H_INLINE int MMRGBHexSimilarToColor(MMRGBHex h1, MMRGBHex h2, float tolerance)
 		uint8_t d1 = RED_FROM_HEX(h1) - RED_FROM_HEX(h2);
 		uint8_t d2 = GREEN_FROM_HEX(h1) - GREEN_FROM_HEX(h2);
 		uint8_t d3 = BLUE_FROM_HEX(h1) - BLUE_FROM_HEX(h2);
-		return sqrtf((d1 * d1) +
-		             (d2 * d2) +
-		             (d3 * d3)) <= (tolerance * 442.0f);
+		return sqrt((d1 * d1) +
+		            (d2 * d2) +
+		            (d3 * d3)) <= (tolerance * 442.0f);
 	}
 }
